Fix variable type to hold virtio-mmio IRQ
The u8 type is not enough to cover all possible values, use u32
instead.

+++ include/kvm/disk-image.h
@@ -58,7 +58,7 @@ struct disk_image_params {
 	bool direct;
 
 	u32 addr;
-	u8 irq;
+	u32 irq;
 };
 
 struct disk_image {
@@ -81,7 +81,7 @@ struct disk_image {
 	int				debug_iodelay;
 
 	u32 addr;
-	u8 irq;
+	u32 irq;
 };
 
 #if 0
